Database schema updates
In order to support tracking write-in votes and no votes in election results, I made it so that there can be candidates not attached to a person that are marked as one of those types. I don't love it, but can't think of anything better at the moment.

Can't remember if it's right to accept null candidates in election_results...

Renamed votes to town_meeting_votes for clarity, and each vote needs to have a person and motion attached to it.